+++ stylis.js
@@ -1,4 +1,4 @@
-/*!
+/*
  *          __        ___     
  *    _____/ /___  __/ (_)____
  *   / ___/ __/ / / / / / ___/
@@ -27,79 +27,106 @@
 	 *
 	 * @example compiler('.class1', 'css...', false);
 	 * 
-	 * @param  {string}  selector
-	 * @param  {string}  styles
-	 * @param  {boolean} nsAnimations
-	 * @param  {boolean} nsKeyframes
+	 * @param  {string}              selector   - i.e `.class` or `#id` or `[attr=id]`
+	 * @param  {string}              styles
+	 * @param  {boolean=}            animations - pass false to prevent prefixing animations, true by default
+	 * @param  {function({string})=} middleware
 	 * @return {string}
 	 */
-	function stylis (selector, styles, nsAnimations, nsKeyframes) {
+	function stylis (selector, styles, animations, middleware) {
 		var prefix = '';
-		var id     = '';
-		var type   = selector.charCodeAt(0) || 0;
+		var namespace = '';
+		var type = selector.charCodeAt(0) || 0;
 
 		// [ attr selector
 		if (type === 91) {
 			// `[data-id=namespace]` -> ['data-id', 'namespace']
 			var attr = selector.substring(1, selector.length-1).split('=');     
-			var char = (id = attr[1]).charCodeAt(0);
+			var char = (namespace = attr[1]).charCodeAt(0);
 
 			// [data-id="namespace"]/[data-id='namespace']
 			// --> "namespace"/'namspace' -> namespace
 			if (char === 34 || char === 39) {
-				id = id.substring(1, id.length-1);
+				namespace = namespace.substring(1, namespace.length-1);
 			}
 
-			prefix = '['+ attr[0] + '=\'' + id +'\']';
+			prefix = '['+ attr[0] + '="' + namespace +'"]';
 		}
 		// `#` `.` `>` id class and descendant selectors
 		else if (type === 35 || type === 46 || type === 62) {
-			id = (prefix = selector).substring(1);
+			namespace = (prefix = selector).substring(1);
 		}
 		// element selector
 		else {
-			id = prefix = selector;
+			namespace = prefix = selector;
 		}
 
-		var keyframeNs  = (nsAnimations === void 0 || nsAnimations === true ) ? id : '';
-		var animationNs = (nsKeyframes === void 0 || nsKeyframes === true ) ? id : '';
+		// animation and keyframe namespace
+		var anim = (animations === void 0 || animations === true) ? namespace : '';
 
-		var output  = '';
-		var line    = '';
-		var blob    = '';
-		var prev    = '';
-		var flat    = '';
+		// has middleware
+		var plugin = middleware !== void 0 && typeof middleware === 'function';
 
-		var len     = styles.length;
+		// variables, i.e $a = 1;
+		var vars;
 
-		var i       = 0;
+		// mixins
+		var mixins;
+		var mixin;
+
+		// buffers
+		var buff = '';
+		var blob = '';
+		var prev = '';
+		var flat = '';
+		var blck = '';
+		var indexOf = 0;
+
+		// position
+		var caret = 0;
+		var depth = 0;
+		var column = 0;
+		var line = 1;
+		var eof = styles.length;
+
+		// context signatures
 		var special = 0;
-		var type    = 0;
-		var close   = 0;
+		var type = 0;
+		var close = 0;
 		var comment = 0;
-		var depth   = 0;
+		var strings = 0;
+
+		var output = '';
 
 		// parse + compile
-		while (i < len) {
-			var code = styles.charCodeAt(i);
+		while (caret < eof) {
+			var code = styles.charCodeAt(caret);
 
 			// {, }, ; characters, parse line by line
-			if (code === 123 || code === 125 || code === 59) {
-				line += styles[i];
+			if (strings === 0 && (code === 123 || code === 125 || code === 59)) {
+				buff += styles[caret];
 
-				var first = line.charCodeAt(0);
+				var first = buff.charCodeAt(0);
 
 				// only trim when the first character is a space ` `
 				if (first === 32) { 
-					first = (line = line.trim()).charCodeAt(0); 
+					first = (buff = buff.trim()).charCodeAt(0); 
 				}
 
-				// default to 0 instead of NaN if there is no second character
-				var second = line.charCodeAt(1) || 0;
+				// default to 0 instead of NaN if there is no second/third character
+				var second = buff.charCodeAt(1) || 0;
+				var third = buff.charCodeAt(2) || 0;
+
+				// middleware
+				plugin && (buff = middleware(0, buff, line, column) || buff);
 
 				// ignore comments
 				if (comment === 2) {
-					line = ''; comment = 0;
+					if (code === 125) {
+						comment = 0;
+					}
+
+					buff = ''; 
 				}
 				// @, special block
 				else if (first === 64) {
@@ -107,41 +134,153 @@
 					if (second === 107 || second === 103) {
 						// k, @keyframes
 						if (second === 107) {
-							blob = line.substring(1, 11) + keyframeNs + line.substring(11);
-							line = '@-webkit-'+blob;
+							blob = buff.substring(1, 11) + anim + buff.substring(11);
+							buff = '@-webkit-'+blob;
 							type = 1;
 						}
 						// g, @global 
 						else {
-							line = '';
+							buff = '';
 						}
 					}
-					// @media `m` character
+					// @media/@mixin `m` character
 					else if (second === 109) {
-						type = 2;
+						if (third === 105) {
+							// first match create mixin store
+							mixins === void 0 && (mixins = {});
+
+							// retrieve mixin identifier
+							blob = (mixin = buff.substring(7, buff.indexOf('{')) + ' ').trim();
+							mixin = mixin.substring(0, mixin.indexOf(' ')).trim();
+
+							// append mixin identifier
+							mixins[mixin] = { key: blob.trim(), body: null };
+
+							type = 3;
+							buff = '';
+							blob = '';
+						}
+						// @media 
+						else {
+							type = 2;
+						}
 					}
 
-					special++;
-				}
-				else {
-					var third = line.charCodeAt(2) || 0;
+					// @include/@import `i` character
+					if (second === 105) {
+						// @include `n` character
+						if (third === 110) {
+							buff = buff.substring(9, buff.length-1);
 
+							indexOf = buff.indexOf('(');
+
+							// function mixins
+							if (indexOf > -1) {
+								// mixin name
+								var name = buff.substring(0, indexOf);
+
+								// mixin data
+								var data = mixins[name];
+
+								// args passed to the mixin
+								var argsPassed = buff.substring(name.length+1, buff.length-1).split(',');
+
+								// args the mixin expects
+								var argsExpected = data.key.replace(name, '').replace(/\(|\)/g, '').trim().split(',');
+								
+								buff = data.body;
+
+								for (var i = 0, length = argsPassed.length; i < length; i++) {
+									var arg = argsExpected[i].trim();	
+
+									// if the mixin has a slot for that arg
+									if (arg !== void 0) {
+										buff = buff.replace(new RegExp('\\'+arg, 'g'), argsPassed[i].trim());
+									}
+								}
+
+								// create block and update styles length
+								styles += buff;
+								eof += buff.length;
+
+								// reset
+								buff = '';
+							}
+							// static mixins
+							else {
+								buff = mixins[buff].body;
+
+								if (depth === 0) {
+									// create block and update styles length
+									styles += buff;
+									eof += buff.length;
+
+									// reset
+									buff = '';
+								}
+							}
+						}
+						// @import `m` character
+						else if (third === 109 && plugin) {
+							// avoid "foo.css"; "foo" screen; "http://foo.com/bar"; url(foo);
+							var match = /@import.*?(["'][^\.\n\r]*?["'];|["'].*\.scss["'])/g.exec(buff);								
+
+							if (match !== null) {
+								// middleware
+								buff = middleware(3, match[1].replace(/['"; ]/g, ''), line, column) || '';
+
+								if (buff) {
+									// create block and update styles length
+									styles += buff;
+									eof += buff.length;
+								}
+
+								buff = '';
+							}
+						}
+					} else {
+						close = -1;
+						special++;
+					}
+				}
+				// $, variable
+				else if (first === 36) {
+					var colon = buff.indexOf(':');
+
+					// first match create variables store 
+					vars === void 0 && (vars = []);
+
+					// push key value pair
+					vars[vars.length] = [buff.substring(0, colon), buff.substring(colon+1, buff.length-1).trim()];
+
+					// reset buffer
+					buff = '';
+				}
+				// property/selector
+				else {
 					// animation: a, n, i characters
 					if (first === 97 && second === 110 && third === 105) {
-						var anims = line.substring(10).split(',');
-						var build = 'animation:';
+						var ninth = buff.charCodeAt(9) || 0;
+						var tenth = buff.charCodeAt(10) || 0;
+
+						var colon = buff.indexOf(':')+1;
+						var build = buff.substring(0, colon);
+						var anims = buff.substring(colon).split(',');
+
+						// -, n
+						var ns = (ninth !== 45 ? anim : ( tenth !== 110 ? '' : anim));
 
 						for (var j = 0, length = anims.length; j < length; j++) {
-							build += (j === 0 ? '' : ',') + animationNs + anims[j].trim();
+							build += (j === 0 ? '' : ',') + ns + anims[j].trim();
 						}
 
 						// vendor prefix
-						line = '-webkit-' + build + build;
+						buff = '-webkit-' + build + build;
 					}
 					// appearance: a, p, p
 					else if (first === 97 && second === 112 && third === 112) {
 						// vendor prefix -webkit- and -moz-
-						line = '-webkit-' + line + '-moz-' + line + line;
+						buff = '-webkit-' + buff + '-moz-' + buff + buff;
 					}
 					// hyphens: h, y, p
 					// user-select: u, s, e
@@ -150,7 +289,7 @@
 						(first === 117 && second === 115 && third === 101)
 					) {
 						// vendor prefix all
-						line = '-webkit-' + line + '-moz-' + line + '-ms-' + line + line;
+						buff = '-webkit-' + buff + '-moz-' + buff + '-ms-' + buff + buff;
 					}
 					// flex: f, l, e
 					// order: o, r, d
@@ -159,18 +298,18 @@
 						(first === 111 && second === 114 && third === 100)
 					) {
 						// vendor prefix only -webkit-
-						line = '-webkit-' + line + line;
+						buff = '-webkit-' + buff + buff;
 					}
 					// transforms & transitions: t, r, a 
 					else if (first === 116 && second === 114 && third === 97) {
 						// vendor prefix -webkit- and -ms- if transform
-						line = '-webkit-' + line + (line.charCodeAt(5) === 102 ? '-ms-' + line : '') + line;
+						buff = '-webkit-' + buff + (buff.charCodeAt(5) === 102 ? '-ms-' + buff : '') + buff;
 					}
 					// display: d, i, s
 					else if (first === 100 && second === 105 && third === 115) {
-						if (line.indexOf('flex') > -1) {
+						if (buff.indexOf('flex') > -1) {
 							// vendor prefix
-							line = 'display:-webkit-flex; display:flex;';
+							buff = 'display:-webkit-flex; display:flex;';
 						}
 					}
 					// { character, selector declaration
@@ -181,25 +320,30 @@
 							// nested selector
 							if (depth === 2) {
 								// discard first character {
-								i++;
+								caret++;
 
 								// inner content of block
 								var inner   = '';
-								var nestSel = line.substring(0, line.length-1).split(',');
+								var nestSel = buff.substring(0, buff.length-1).split(',');
 								var prevSel = prev.substring(0, prev.length-1).split(',');
 
 								// keep track of opening `{` and `}` occurrences
 								var counter = 1;
 
 								// travel to the end of the block
-								while (i < len) {
-									var char = styles.charCodeAt(i);
+								while (caret < eof) {
+									var char = styles.charCodeAt(caret);
+
 									// {, }, nested blocks may have nested blocks
 									char === 123 ? counter++ : char === 125 && counter--;
+
 									// break when the has ended
-									if (counter === 0) break;
+									if (counter === 0) {
+										break;
+									}
+
 									// build content of nested block
-									inner += styles[i++];
+									inner += styles[caret++];
 								}
 
 								// handle multiple selectors: h1, h2 { div, h4 {} } should generate
@@ -207,6 +351,7 @@
 								for (var j = 0, length = prevSel.length; j < length; j++) {
 									// extract value, prep index for reuse
 									var val = prevSel[j]; prevSel[j] = '';
+
 									// since there can also be multiple nested selectors
 									for (var k = 0, l = nestSel.length; k < l; k++) {
 										prevSel[j] += (
@@ -217,17 +362,18 @@
 								}
 
 								// create block and update styles length
-								len += (styles += (prevSel.join(',') + '{'+inner+'}').replace(/&| +&/g, '')).length;
+								// the new line is to avoid conflicts when the last line is a // line comments
+								eof += (styles += ('\n' + prevSel.join(',') + '{'+inner+'}').replace(/&| +&/g, '')).length;
 
 								// clear current line, to avoid add block elements to the normal flow
-								line = '';
+								buff = '';
 
 								// decreament depth
 								depth--;
 							}
 							// top-level selector
 							else {
-								var split = line.split(',');
+								var split = buff.split(',');
 								var build = '';
 
 								// prefix multiple selectors with namesapces
@@ -305,7 +451,8 @@
 									build += j === 0 ? selector : ',' + selector;
 								}
 
-								prev = line = build;
+								// cache current selector
+								prev = (buff = build);
 							}
 						}
 					}
@@ -320,68 +467,142 @@
 						code === 125 ? close++ : (code === 123 && close !== 0 && close--);
 
 						// closing tag
-						if (close === 2) {
+						if (close === 0) {
 							// @global
 							if (type === 0) {
-								line = '';
+								buff = '';
 							}
 							// @keyframes 
 							else if (type === 1) {
 								// vendor prefix
-								line = '}@'+blob+'}';
-								// reset blob
+								buff = '}@'+blob+'}';
+
+								// reset
 								blob = '';
 							}
 							// @media
 							else if (type === 2) {
-								blob.length !== 0 && (line = prefix + ' {'+blob+'}' + line);
-								// reset blob
+								blob.length !== 0 && (buff = prefix + ' {'+blob+'}' + buff);
+
+								// reset
+								blob = '';
+							}
+							// @mixin
+							else if (type === 3) {
+								// append body of mixin
+								mixins[mixin].body = blob;
+
+								// reset
+								mixin = '';
+								buff = '';
 								blob = '';
 							}
 
 							// reset flags
-							type = 0; close--; special--;
+							type = 0;
+							close--;
+							special--;
 						}
-						// @keyframes 
-						else if (type === 1) {
-							blob += line;
+						// @keyframes, @mixin
+						else if (type === 1 || type === 3) {
+							blob += buff;
+
+							if (type === 3) {
+								buff = '';
+							}
 						}
 						// @media flat context
 						else if (type === 2 && depth === 0 && code !== 125) {
-							blob += line; line = '';
+							blob += buff; 
+							buff = '';
 						}
 					}
 					// flat context
 					else if (depth === 0 && code !== 125) {
-						flat += line; line = '';
+						flat += buff; 
+						buff = '';
 					}
 				}
 
-				// add line to output, reset line buffer and comment signal
-				output += line; line = ''; comment = 0;
+				// append line to blck buffer
+				blck += buff; 
+
+				// reset line buffer
+				buff = ''; 
+
+				// add blck buffer to output, 
+				if (code === 125 && type === 0 && comment === 0) {
+					// middleware
+					plugin && blck.length !== 0 && (blck = middleware(1, blck, line, column) || blck);
+
+					// append blck buffer
+					output += blck;
+
+					// reset blck buffer
+					blck = '';
+				}
 			}
 			// build line by line
 			else {
-				// \r, \n, ignore line and block comments
-				if (comment === 2 && (code === 13 || code === 10)) {
-					line = ''; comment = 0;
+				// \r, \n, new lines
+				if (code === 13 || code === 10) {
+					// ignore line and block comments
+					if (comment === 2) {
+						buff = ''; 
+						comment = 0;
+					}
+
+					column = 0;
+					line++;
 				}
-				// not `\t`, `\r`, `\n` characters
-				else if (code !== 9 && code !== 13 && code !== 10) {
-					// / line comment signal
-					code === 47 && comment < 2 && comment++;
+				// not `\t` tab character
+				else if (code !== 9) {
+					// " character
+					if (code === 34) {
+						// exit string " context / enter string context
+						strings = strings === 34 ? 0 : 34;
+					}
+					// ' character
+					else if (code === 39) {
+						// exit string ' context / enter string context
+						strings = strings === 39 ? 0 : 39;
+					}
+					// / character
+					else if (code === 47) {
+						code === 47 && comment < 2 && comment++;
+					}
 
 					// build line buffer
-					line += styles[i];
+					buff += styles[caret];
 				}
 			}
 
-			// next character
-			i++; 
+			// move caret position
+			caret++;
+
+			// move column position
+			column++;
 		}
 
-		// if there is flat css, append
-		return output + (flat.length === 0 ? '' : prefix + ' {' + flat + '}');
+		// has flat css 
+		if (flat.length !== 0) {
+			flat = prefix + ' {' + flat + '}';
+
+			// middleware
+			plugin && (flat = middleware(2, flat, line, column) || flat);
+
+			// append flat css
+			output += flat;
+		}
+
+		// replace variables
+		if (vars !== void 0) {
+			for (var i = 0, l = vars.length; i < l; i++) {	
+				output = output.replace(new RegExp('\\'+vars[i][0], 'g'), vars[i][1]);
+			}
+		}
+
+		return output;
 	}
 
 	return stylis;