+++ tests/index.js
@@ -30,6 +30,10 @@
  */
 var tests = {
 	'flat': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'flat',
 		sample: `
 			color: 20px;
@@ -39,6 +43,10 @@ var tests = {
 		expected: `.user {color: 20px;font-size: 20px;-webkit-transition: all;transition: all;}`
 	},
 	'namespace': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'namespace',
 		sample: `
 			{
@@ -48,6 +56,10 @@ var tests = {
 		expected: `.user {color: blue;}`
 	},
 	'globals': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: '@global/:global',
 		sample: `
 			@global {
@@ -62,6 +74,10 @@ var tests = {
 		expected: 'body {background: yellow;}body {background: yellow;}'
 	},
 	'comment': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'comments',
 		sample: `
 			// line comment
@@ -75,6 +91,10 @@ var tests = {
 		expected: ''
 	},
 	'&': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: '&',
 		sample: `
 			&{
@@ -84,6 +104,10 @@ var tests = {
 		expected: '.user{color: blue;}'
 	},
 	'&:before': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: '&:before',
 		sample: `
 			&:before{
@@ -93,6 +117,10 @@ var tests = {
 		expected: '.user:before{color: blue;}'
 	},
 	'@media': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: '@media',
 		sample: `
 			@media (max-width: 600px) {
@@ -104,6 +132,10 @@ var tests = {
 		expected: `@media (max-width: 600px) {.user h1 {color: red;}.user {color: red;display: none;}}`
 	},
 	'multiple selectors': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'multiple selectors',
 		sample: `
 			span, h1 {
@@ -116,6 +148,10 @@ var tests = {
 		expected: `.user span,.user h1 {color:red;}.user h1,.user:after,.user:before {color:red;}`
 	},
 	'prefixer': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'prefixer',
 		sample: `
 			& {
@@ -124,7 +160,44 @@ var tests = {
 		`,
 		expected: `.user {-webkit-transform: rotate(30deg);-ms-transform: rotate(30deg);transform: rotate(30deg);}`
 	},
+	'animations': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
+		name: 'animations',
+		sample: `
+			span {
+      			animation-duration: 0.6s;
+      			animation-name: slidein;
+      			animation-iteration-count: infinite;
+		    }
+		`,
+		expected: `.user span {-webkit-animation-duration:0.6s;animation-duration:0.6s;-webkit-animation-name:userslidein;`+
+			`animation-name:userslidein;-webkit-animation-iteration-count:infinite;animation-iteration-count:infinite;}`
+	},
+	'animations disabled prefix': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false,
+			namespaceAnimations: false
+		},
+		name: 'animations disable namespace',
+		sample: `
+			span {
+      			animation-duration: 0.6s;
+      			animation-name: slidein;
+      			animation-iteration-count: infinite;
+		    }
+		`,
+		expected: `.user span {-webkit-animation-duration:0.6s;animation-duration:0.6s;-webkit-animation-name:slidein;`+
+			`animation-name:slidein;-webkit-animation-iteration-count:infinite;animation-iteration-count:infinite;}`
+	},
 	'keyframes': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'keyframes',
 		sample: `
 			&{
@@ -142,7 +215,34 @@ var tests = {
 			`@keyframes userslidein `+
 			`{to {-webkit-transform: translate(20px);-ms-transform: translate(20px);transform: translate(20px);}}`
 	},
+	'keyframes': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false,
+			namespaceAnimations: false
+		},
+		name: 'keyframes disable namespace',
+		sample: `
+			&{
+				animation: slidein 3s ease infinite;
+			}
+			@keyframes slidein {
+				to { transform: translate(20px); }
+			}
+		`,
+		expected: `.user{-webkit-animation:slidein 3s ease infinite;animation:slidein 3s ease infinite;}`+
+
+			`@-webkit-keyframes slidein `+
+			`{to {-webkit-transform: translate(20px);-ms-transform: translate(20px);transform: translate(20px);}}`+
+
+			`@keyframes slidein `+
+			`{to {-webkit-transform: translate(20px);-ms-transform: translate(20px);transform: translate(20px);}}`
+	},
 	':host': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: ':host',
 		sample: `
 			color: red;
@@ -158,6 +258,10 @@ var tests = {
 			'body .user {color: red;}.user:root {color: red;}.user {color: red;}'
 	},
 	'[title="a,b"]': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: '[title="a,b"]',
 		sample: `
 			[title="a,b,c, something"], h1 {
@@ -167,6 +271,10 @@ var tests = {
 		expected: `.user [title="a,b,c, something"],.user h1 {color: red}`
 	},
 	'nested': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
 		name: 'nested',
 		sample: `
 			h1, div {
@@ -185,6 +293,32 @@ var tests = {
 		`,
 		expected: '.user h1,.user div {color: red;color: blue;}.user h1 h2,.user h1:before,.user div h2,'+
 		'.user div:before{color: red;}.user h1 header,.user div header{font-size: 12px;}'
+	},
+	'variables': {
+		options: {
+			disableWebComponents: false,
+			disableFlat: false
+		},
+		name: 'variables',
+		sample: `
+			$foo: 20px;
+
+			width: $foo;
+
+			& {
+				margin: $foo;
+			}
+		`,
+		expected: '.user {margin: 20px;}.user {width: 20px;}'
+	},
+	'strings': {
+		name: 'strings',
+		sample: `
+			.foo:before {
+		  		content: ".hello {world}"
+			}
+		`,
+		expected: `.user .foo:before {content: ".hello {world}"}`
 	}
 };
 
@@ -219,8 +353,16 @@ function run (tests) {
 		var name = test.name.trim();
 		var sample = test.sample.trim();
 		var expected = test.expected.trim();
+		var options = test.options || {};
 
-		var result = stylis('.user', sample, true, true);
+		var result = stylis(
+			'.user', 
+			sample, 
+			options.namespaceAnimations, 
+			!options.disableFlat, 
+			!options.disableWebComponents, 
+			options.middleware
+		);
 
 		(result === expected ? passed : failed).push(name);